Enable syndication in an event handler.
We used to do this in initializeArchetype, but in Plone 4.2 this does not always work.
Archetypes creates objects a bit too often in portal_factory and in one of those times it could go wrong for some reason.
In the end it did work, but for every Collection you would get a traceback in the error log on creation.